evalutation db created completed but not updated completed

+++ src/main/java/com/site/restauranttier/etc/JsonData.java
@@ -8,7 +8,7 @@ import java.util.List;
 @Getter
 @Setter
 public class JsonData {
-    int starRating;
+    Double starRating;
     int restaurantId;
-    List<Integer> barRatings;
+    List<Double> barRatings;
 }
